```
benchstat -delta-test=none ../_dev/thanos/2021/receive2/impr5-go1.16.3.txt ../_dev/thanos/2021/receive2/errimpr1-go1.16.3.txt
name                                                                                old time/op    new time/op    delta
HandlerReceiveHTTP/typical_labels_under_1KB,_500_of_them/OK-12                        1.22ms ± 0%    1.05ms ± 0%  -13.97%
HandlerReceiveHTTP/typical_labels_under_1KB,_500_of_them/conflict_errors-12           5.84ms ± 0%    1.26ms ± 0%  -78.46%
HandlerReceiveHTTP/typical_labels_under_1KB,_5000_of_them/OK-12                       12.8ms ± 0%    11.3ms ± 0%  -11.54%
HandlerReceiveHTTP/typical_labels_under_1KB,_5000_of_them/conflict_errors-12          59.4ms ± 0%    12.8ms ± 0%  -78.39%
HandlerReceiveHTTP/extremely_large_label_value_10MB,_10_of_them/OK-12                  109ms ± 0%      97ms ± 0%  -11.16%
HandlerReceiveHTTP/extremely_large_label_value_10MB,_10_of_them/conflict_errors-12     1.44s ± 0%     0.10s ± 0%  -93.25%

name                                                                                old alloc/op   new alloc/op   delta
HandlerReceiveHTTP/typical_labels_under_1KB,_500_of_them/OK-12                        1.20MB ± 0%    1.20MB ± 0%   -0.69%
HandlerReceiveHTTP/typical_labels_under_1KB,_500_of_them/conflict_errors-12           4.34MB ± 0%    1.36MB ± 0%  -68.54%
HandlerReceiveHTTP/typical_labels_under_1KB,_5000_of_them/OK-12                       13.3MB ± 0%    13.2MB ± 0%   -0.67%
HandlerReceiveHTTP/typical_labels_under_1KB,_5000_of_them/conflict_errors-12          44.6MB ± 0%    14.9MB ± 0%  -66.67%
HandlerReceiveHTTP/extremely_large_label_value_10MB,_10_of_them/OK-12                  120MB ± 0%     120MB ± 0%   -0.00%
HandlerReceiveHTTP/extremely_large_label_value_10MB,_10_of_them/conflict_errors-12     592MB ± 0%     120MB ± 0%  -79.77%

name                                                                                old allocs/op  new allocs/op  delta
HandlerReceiveHTTP/typical_labels_under_1KB,_500_of_them/OK-12                         5.10k ± 0%     4.59k ± 0%   -9.83%
HandlerReceiveHTTP/typical_labels_under_1KB,_500_of_them/conflict_errors-12            25.1k ± 0%      7.6k ± 0%  -69.67%
HandlerReceiveHTTP/typical_labels_under_1KB,_5000_of_them/OK-12                        50.4k ± 0%     45.4k ± 0%   -9.93%
HandlerReceiveHTTP/typical_labels_under_1KB,_5000_of_them/conflict_errors-12            250k ± 0%       75k ± 0%  -69.96%
HandlerReceiveHTTP/extremely_large_label_value_10MB,_10_of_them/OK-12                   89.0 ± 0%      78.0 ± 0%  -12.36%
HandlerReceiveHTTP/extremely_large_label_value_10MB,_10_of_them/conflict_errors-12       368 ± 0%       201 ± 0%  -45.38%
```

Signed-off-by: Bartlomiej Plotka <bwplotka@gmail.com>

<!--
    Keep PR title verbose enough and add prefix telling
    about what components it touches e.g "query:" or ".*:"
-->

<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [ ] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

## Changes

<!-- Enumerate changes you made -->

## Verification

<!-- How you tested it? How do you know it works? -->