@bwplotka 2021-04-20
With this, it's as minimal as it can be I think